+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 - Added TEMPLATE view that allow to render your template with custom Handlebars template.
 - Improved syntax highlighting - now lines with errors will get highlighted with appropriate colour to indicate the issue
+- Added @path, @fileName and @extension variables you can use inside your SQL alongside other Frontmatter properties
 
 Fixes:
 - Fixed parsing arguments to the `file` function. Now parameters with symbols like `[]*` should work properly (i.e. JSONPath arguments)